requirements.txt
apache-airflow-providers-mysql==4.0.2
